edit menu made dynamic now that APIs works
